Combine Solution with NoxSolution and updated tests accordingly
