+++ tests/Nox.Generator.Tests/Domain/QueryTests.cs
@@ -30,8 +30,7 @@ public class QueryTests : IClassFixture<GeneratorFixture>
             "Application.Factories.CountryFactory.g.cs"
         };
 
-        _fixture
-            .GenerateSourceCodeFor(sources)
+        GeneratorFixture.GenerateSourceCodeFor(sources)
             .AssertOutputResult()
             .AssertFileCount(8, filesShouldExist)
             .AssertContent()
