+++ tests/Nox.Generator.Tests/Ui/UiGeneratorTests.cs
@@ -21,8 +21,7 @@ public class UiGeneratorTests : IClassFixture<GeneratorFixture>
             $"./{path}dto.solution.nox.yaml"
         };
 
-        _fixture
-            .GenerateSourceCodeFor(sources)
+        GeneratorFixture.GenerateSourceCodeFor(sources)
             .AssertOutputResult()
             .AssertContent()
             .SourceContains("Generator.g.cs", "SUCCESS");
@@ -39,8 +38,7 @@ public class UiGeneratorTests : IClassFixture<GeneratorFixture>
             $"./{path}dto.solution.nox.yaml"
         };
 
-        _fixture
-            .GenerateSourceCodeFor(sources)
+        GeneratorFixture.GenerateSourceCodeFor(sources)
             .AssertOutputResult()
             .AssertContent()
             .SourceContains("Generator.g.cs", "Error");
@@ -48,7 +46,6 @@ public class UiGeneratorTests : IClassFixture<GeneratorFixture>
         // This trick is necessary because broken configuration impacts on test Generated_Files_Should_Be_Compiled_Successfully
         // Seems that driver.RunGenerators() does not release generated resources on time.
         // Another fix is to set Thread.Sleep(2000) in test Generated_Files_Should_Be_Compiled_Successfully.
-        _fixture
-            .GenerateSourceCodeFor(new[] { $"./{path}generator.nox.yaml" });
+        GeneratorFixture.GenerateSourceCodeFor(new[] { $"./{path}generator.nox.yaml" });
     }
 }
